trying to configure weatherItems
